Moved duels to a separate channel

+++ index.js
@@ -124,9 +124,13 @@ client.on('message', msg => {
                                 c.foo(args, msg, client)
                             })
                     }
-                    else if(msg.channel.id == constants.channels.general && !c.allowedInGeneral)
+                    else if(args[0] == 'duel' && msg.channel.id == constants.channels.duels) {
                         msg.delete()
-                    else {
+                            .then(() => {
+                                c.foo(args, msg, client)
+                            })
+                    }
+                    else if(!args[0] == 'duel') {
                         msg.delete()
                             .then(() => {
                                 c.foo(args, msg, client)
